Fixup more signed/unsigned comparisons.
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk@3790 06827809-a52a-0410-b366-d66718629ded

+++ snmplib/snmpusm.c
@@ -1443,7 +1443,7 @@ usm_check_and_update_timeliness(
 	/*
 	 * This is a local reference.
 	 */
-	if ( secEngineIDLen == myIDLength
+	if ( (int)secEngineIDLen == myIDLength
 		&& memcmp (secEngineID, myID, myIDLength) == 0 )
 	{
 		u_int time_difference = myTime > time_uint ?
